Switch to use GitPython
The primary purpose for doing this is easy access to the
most recent commit for a revision, but in general it seems
a bit tidier.

Except on front: you can't seem to set the author and committer
on a commit in any straightforward fashion. It's probably possible
but some hopping and skipping required.

The long game here is be able to easily get to the first revision
of a tiddler.